profile restructure
