+++ src/Core/Application/Identity/Users/Profile/ResetPasswordRequest.cs
@@ -1,4 +1,4 @@
-namespace FSH.WebApi.Application.Identity.Users;
+namespace FSH.WebApi.Application.Identity.Users.Profile;
 
 public class ResetPasswordRequest
 {
