+++ src/Core/Application/Identity/Users/Profile/UpdateProfileRequestValidator.cs
@@ -1,4 +1,4 @@
-namespace FSH.WebApi.Application.Identity.Users;
+namespace FSH.WebApi.Application.Identity.Users.Profile;
 
 public class UpdateProfileRequestValidator : CustomValidator<UpdateProfileRequest>
 {
